refactor: Update logout button to use API for logout functionality
This commit updates the logout button component to use the API for the logout functionality instead of directly calling the logout function from the useUserStore hook. The commented out code that made a POST request to "/logout" has been replaced with the call to the logout function. After logging out, the user is redirected to the login page ("/").