refactor: Add imageUrl field to post schema
This commit adds the imageUrl field to the post schema in the server/models/post.js file. The imageUrl field is of type String and is required, indicating that an image URL is mandatory for a post. This update enhances the functionality of the post schema by allowing posts to have associated images.