refactor: Update main container height in MarketingLayout component
This commit updates the main container height in the MarketingLayout component of the client app. The previous height was set to "pt-56" which caused unnecessary vertical space. The new height is set to "h-full" to ensure the container takes up the full height of the screen. This change improves the layout and visual appearance of the page.